First main commit of the files.
The code runs when running main.py, it will ask questions that need to be answered in the console.

First, you will have to get your own OpenAI API key, as well as various keys for use with the Spotify API, which you can mostly get from the Spotify for Developers site when making an application.
You can use these with the commands in the commands_to_run file to get everything working.

Annoingly, this current setup then requires specifying a redirect URI which you will be redirected to multiple times, each time just copy and paste the link you are sent to into the console.

Next steps will involve finding a cleaner approach, improving the prompting, and adding the auto generation of an icon for the playlist.